backend/requirements.txt
langchain
fastapi
uvicorn
pydantic
pytest
flake8
langchain-google-genai
python-pptx
Pillow 
pydantic-settings 